pyproject.toml
[project]
name = "wechat-group-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "suiseiseki-2016"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "uiautomation (>=2.0.20,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pywin32 (>=308,<309)",
    "psutil (>=7.0.0,<8.0.0)",
    "pyperclip (>=1.9.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.63.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
